# docs/swagger.yaml
basePath: /
definitions:
...
  models.Report:
    properties:
      createdAt:
        type: string
      id:
        type: string
      postId:
        type: string
      reason:
        $ref: '#/definitions/models.ReportReason'
      reportedBy:
        type: string
    type: object
  models.ReportCreate:
    properties:
      reason:
        $ref: '#/definitions/models.ReportReason'
    required:
    - reason
    type: object
  models.ReportReason:
    enum:
    - DISLIKE
    - HARASSMENT
    - SELF_HARM
    - VIOLENCE
    - RESTRICTED_ITEMS
    - NUDITY
    - SCAM
    - MISINFORMATION
    - ILLEGAL_CONTENT
    type: string
    x-enum-varnames:
    - DISLIKE
    - HARASSMENT
    - SELF_HARM
    - VIOLENCE
    - RESTRICTED_ITEMS
    - NUDITY
    - SCAM
    - MISINFORMATION
    - ILLEGAL_CONTENT
...
host: localhost:8090
info:
  contact: {}
  description: API pour le projet PEC2 Backend
  title: API PEC2 Backend
  version: "1.0"
paths:
...
  /posts/{id}/report:
    post:
      consumes:
      - application/json
      description: Report a post for inappropriate content
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Report reason
        in: body
        name: report
        required: true
        schema:
          $ref: '#/definitions/models.ReportCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Report'
        "400":
          description: 'error: Invalid input'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Post not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Error message'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Report a post
      tags:
      - posts
  /posts/reports:
    get:
      description: Get all reports with optional filtering
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Report'
            type: array
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Error message'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all reports (Admin only)
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: 'Entrez le JWT avec le préfixe Bearer: Bearer <JWT>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"